app/requirements.txt
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==5.0.1
bs4==0.0.1
business-duration==0.66
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-auth-adfs==1.11.6
django-autocomplete-light==3.9.7
django-bleach==3.1.0
django-bootstrap-datepicker-plus==5.0.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-constance==3.1.0
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==23.2
django-fsm==2.8.1
django-guardian==2.4.0
django-ical==1.9.2
django-impersonate==1.9.1
django-import-export==3.2.0
django-location-field==2.7.2
django-model-utils==4.3.1
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-picklefield==3.1
django-recurrence==1.11.1
django-simple-history==3.3.0
django-simple-menu==2.1.2.post1
django-sql-explorer==3.2
django-timezone-field==5.1
django-tinymce==3.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==18.13.0
faker-microservice==2.0.0
graphviz==0.20.1
holidays==0.33
icalendar==5.0.7
idna==3.4
kombu==5.3.1
lxml==4.9.2
MarkupPy==1.14
mysqlclient==2.2.0
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
phonenumbers==8.13.15
Pillow==10.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
pydantic==2.1.0
pydantic_core==2.4.0
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sentry-sdk==1.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
